Updated v3 with "removed tree handling".
All columns should be "tree independent", so the COL_INFO stuff in the switch
statement shouldn't be under a tree either.  In the interest of simplicity I
just removed the tree check entirely.  You could add it back for individual
cases in the switch statement (AFTER the COL_INFO calls), but I don't think the
performance benefit is necessarily worth the effort.

Also, has this been fuzztested?  I don't think "fail" cases have been properly
accounted for, like when tvb_find_guint8() returns -1 to 'offset' (which is a
guint32).  See packet-bitcoin.c for examples as it has to handle the same
issue.