Fix signed/insigned mismatch in ErlDP dissector
This part isn't well documented so we have to carefully read docs. Although the
common sense tells us that a length can't be negative, we'd love to have it
stated clearly. Unfortunately instead of explicitly telling us which fields has
which signedness the Erlang authors either forget to do that completely or told
us implicitly - by stating that some given field contains an integer value and
later defining a value range for that. For example "The range is 0 <=
AtomCacheReferenceIndex < 255".

See these links for further investigation:

* http://www.erlang.org/doc/apps/erts/erl_ext_dist.html
* https://github.com/matwey/pybeam/blob/master/pybeam/eetf_construct.py
* https://github.com/therve/twotp/tree/master/twotp
* https://github.com/goerlang/etf

Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>