Properly fill information about the interfaces
Properly fill information about the interfaces while merging. This fixes
Bug: 8795, Bug: 10308.

We don't know if this number will be the same for all captures to merge.
We have to hope that they will match.

Just a bit more explanation. If we capture dumps from multiple specified
interfaces it's not possible to merge them later with mergecap. While
creating a file to merge mergecap initializes it as if it has to dump a
traffic from a single interface.

For example dump something with the following command (both br0 and
virbr0 interfaces should active send/receive traffic of course):

./dumpcap -t -Z none -i br0 -i virbr0 -w ~/all.log

And later try to merge it with:

./mergecap -v -w ~/all2.log ~/all.log

You'll get a "mergecap: Error writing to outfile: Internal error" as
soon as mergecap finds a packet from interface #1.

Change-Id: I8c1e217a81421ab4091376790b72c93abe937afa
Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>