Allow redefining all ports for RADIUS
RADIUS configuration sometimes uses more ports - for example, one for
authentication, another one for accounting. Sometimes it uses the entire
port ranges. In case of FreeRADIUS 2.x.x server it might look like this:

...
listen {
        type = auth
        ipaddr = *
        port = 13812
}
listen {
        type = acct
        ipaddr = *
        port = 13813
}
...

Unfortunately we allow only one port to be redefined, not more. So it
forces a person who's analyzing a traffic from such a RADIUS server
manually select "Decode as" every time for each port.

It was requested at least once to lift this limitation:

* https://ask.wireshark.org/questions/2189/decode-multiple-ports-as-radius

So let's fix it!

With this commit it's possible to set a port ranges for RADIUS dissector
to handle. An example (default) configuration looks like (see
~/.wireshark/preferences):

radius.ports: 1645,1646,1700,1812,1813,3799

Old "alternate_port" preference is marked as obsolete. It won't be shown
to a user but it will still be used if exists (remained from a previous
installations).

*Ver. 2*:
Old alternate_port value is copied to the ports range, thus making
transition even more smooth.

Change-Id: Ibdd6f4f9fa1e0ac186147cec380bbfc62d509b17
Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>