Patch fixing the wireshark autoconf macros.
When configuring a project using a wireshark detection macro on a 64-bit
system, the detection fails because a configuration test program
crashes.

epan/epan.h include is missing in configuration test program. This is
needed to define epan_get_version() as char *. Failure to doing so uses
a 32-bit integer as a string address --> segfault.